Reorganizando metodos repetidos
