+++ api/tests/auth_views/test_jobcore_invite.py
@@ -4,10 +4,11 @@ import json
 from mock import patch
 from django.urls import reverse_lazy
 from django.apps import apps
+from api.tests.mixins import WithMakeUser, WithMakeShift
 
 
 @override_settings(STATICFILES_STORAGE=None)
-class JobcoreInviteTestSuite(TestCase):
+class JobcoreInviteTestSuite(TestCase, WithMakeUser, WithMakeShift):
     """
     Endpoint tests for login
     """
@@ -41,50 +42,6 @@ class JobcoreInviteTestSuite(TestCase):
             'api.JobCoreInvite',
             sender=self.test_profile_employer,
             shift=self.test_shift)
-
-    def _make_shift(
-            self, employer):
-        venue = mixer.blend('api.Venue', employer=employer)
-        position = mixer.blend('api.Position')
-
-        shift = mixer.blend(
-            'api.Shift',
-            venue=venue,
-            position=position,
-            employer=employer)
-
-        return shift, venue, position
-
-    def _make_user(
-            self, kind, userkwargs={}, employexkwargs={}, profilekwargs={}):
-
-        if kind not in ['employee', 'employer']:
-            raise RuntimeError('Do you know what are you doing?')
-
-        user = mixer.blend('auth.User', **userkwargs)
-        user.set_password('pass1234')
-        user.save()
-
-        emptype = 'api.Employee' if kind == 'employee' else 'api.Employer'
-
-        if kind == 'employee':
-            employexkwargs.update({
-                'user': user
-            })
-
-        emp = mixer.blend(emptype, **employexkwargs)
-        emp.save()
-
-        profilekwargs = profilekwargs.copy()
-        profilekwargs.update({
-            'user': user,
-            kind: emp,
-        })
-
-        profile = mixer.blend('api.Profile', **profilekwargs)
-        profile.save()
-
-        return user, emp, profile
 
     def test_list_jobcore_invites(self):
         """
