+++ api/tests/public_views/test_user_validation_email.py
@@ -1,41 +1,29 @@
 from django.test import TestCase, override_settings
-# from unittest import expectedFailure
-from mixer.backend.django import mixer
 from django.urls.base import reverse_lazy
 from rest_framework_jwt.settings import api_settings
+from api.tests.mixins import WithMakeUser
 
 jwt_payload_handler = api_settings.JWT_PAYLOAD_HANDLER
 jwt_encode_handler = api_settings.JWT_ENCODE_HANDLER
 
 
 @override_settings(STATICFILES_STORAGE=None)
-class UserValidationEmailTestSuite(TestCase):
+class UserValidationEmailTestSuite(TestCase, WithMakeUser):
     """
     Endpoint tests for password reset
     """
     USER_VALID_URL = reverse_lazy('api:validate-email')
 
     def setUp(self):
-        self.test_user = self._make_user_with_profile(
-            username='test_user',
-            email='test_user@testdoma.in',
-            is_active=True,
+        self.test_user, *_ = self._make_user(
+            'employee',
+            userkwargs=dict(
+                username='test_user',
+                email='test_user@testdoma.in',
+                is_active=True,
+            )
         )
 
-    def _make_user_with_profile(self, **kwargs):
-        test_user = mixer.blend(
-            'auth.User',
-            **kwargs
-            )
-
-        test_user.set_password('pass1234')
-        test_user.save()
-
-        test_profile = mixer.blend('api.Profile', user=test_user)
-        test_profile.save()
-        return test_user
-
-    # @expectedFailure
     def test_with_bad_token(self):
         """
         Try to reach the view with a bad token
@@ -55,7 +43,6 @@ class UserValidationEmailTestSuite(TestCase):
             400,
             'It should return an error response')
 
-    # @expectedFailure
     def test_reset_kind_of_bad_token(self):
         """
         Try to reach the form with a bad token, good shape, bad data
@@ -102,7 +89,6 @@ class UserValidationEmailTestSuite(TestCase):
             200,
             'It should return an error response')
 
-    # @expectedFailure
     def test_revalidate(self):
         """
         Try to revalidate user
